# internal/oas/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"net/http"
)

func (s *Server) notFound(w http.ResponseWriter, r *http.Request) {
	s.cfg.NotFound(w, r)
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	if len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}
	// Static code generated router with unwrapped path search.
	switch r.Method {
	case "DELETE":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet/"
			if l := len("/pet/"); len(elem) >= l && elem[0:l] == "/pet/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "petId"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				// Leaf: DeletePet
				s.handleDeletePetRequest([1]string{
					args[0],
				}, w, r)

				return
			}
		}
	case "GET":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet/"
			if l := len("/pet/"); len(elem) >= l && elem[0:l] == "/pet/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "petId"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				// Leaf: GetPetById
				s.handleGetPetByIdRequest([1]string{
					args[0],
				}, w, r)

				return
			}
		}
	case "POST":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet"
			if l := len("/pet"); len(elem) >= l && elem[0:l] == "/pet" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				s.handleAddPetRequest([0]string{}, w, r)

				return
			}
			switch elem[0] {
			case '/': // Prefix: "/"
				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "petId"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf: UpdatePet
					s.handleUpdatePetRequest([1]string{
						args[0],
					}, w, r)

					return
				}
			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name  string
	count int
	args  [1]string
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.name
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
func (s *Server) FindRoute(method, path string) (r Route, _ bool) {
	var (
		args = [1]string{}
		elem = path
	)
	r.args = args
	if elem == "" {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch method {
	case "DELETE":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet/"
			if l := len("/pet/"); len(elem) >= l && elem[0:l] == "/pet/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "petId"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				// Leaf: DeletePet
				r.name = "DeletePet"
				r.args = args
				r.count = 1
				return r, true
			}
		}
	case "GET":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet/"
			if l := len("/pet/"); len(elem) >= l && elem[0:l] == "/pet/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "petId"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				// Leaf: GetPetById
				r.name = "GetPetById"
				r.args = args
				r.count = 1
				return r, true
			}
		}
	case "POST":
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/pet"
			if l := len("/pet"); len(elem) >= l && elem[0:l] == "/pet" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				r.name = "AddPet"
				r.args = args
				r.count = 0
				return r, true
			}
			switch elem[0] {
			case '/': // Prefix: "/"
				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "petId"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf: UpdatePet
					r.name = "UpdatePet"
					r.args = args
					r.count = 1
					return r, true
				}
			}
		}
	}
	return r, false
}